requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
Jinja2==3.1.5
pycryptodome==3.21.0
